+++ example/android/app/src/main/kotlin/com/zerone/zo_screenshot_example/MainActivity.kt
@@ -2,4 +2,4 @@ package com.zerone.zo_screenshot_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()